refactor: % -> px, 줄 띄워쓰기, Props 불러오는 위치 수정
