+++ Frontend/src/components/complete/Summary/Summary.style.tsx
@@ -38,7 +38,7 @@ export const TotalPrice = styled.div`
   font-size: 34px;
   font-style: normal;
   font-weight: 500;
-  line-height: 130%; /* 44.2px */
+  line-height: 44.2px;
   letter-spacing: -1.02px;
 `;
