+++ Frontend/src/constants/index.ts
@@ -1,1 +1,5 @@
 export const DEFAULT_PRICE: number = 43460000;
+
+export const GUIDE_MAX_STEP: number = 3;
+
+export const GUIDE_MAX_TAG_NUM: number = 3;
